requirements.txt
alabaster==1.0.0
asgiref==3.8.1
babel==2.17.0
certifi==2025.6.15
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
Django==5.1.11
django-crispy-forms==2.4
docutils==0.21.2
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
Pygments==2.19.2
requests==2.32.4
roman-numerals-py==3.1.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
urllib3==2.5.0